Avoid multiplication overflow in SKP_SMLALBB
http://findbugs.sourceforge.net/bugDescriptions.html#ICAST_INTEGER_MULTIPLY_CAST_TO_LONG

+++ src/org/jitsi/impl/neomedia/codec/audio/silk/SigProcFIX.java
@@ -141,7 +141,7 @@ public class SigProcFIX
      */
     static long SKP_SMLALBB(long a64, short b16, short c16)
     {
-        return (a64) + ((b16) * (c16));
+        return (a64) + ((long)(b16) * (long)(c16));
     }
 
     /**
